fix swap mode
- fix BSS handling in swap mode
- Split swap features in several files
- Reuse SWAP interface from SDK
- Add a modal on Stax startup in SWAP mode
- Return signing status to Exchange after a SWAP completion
- Remove global hdPath usage in CHECK_ADDRESS
- Improve security against double signing attempts
